# docs/public/openapi.json
{
  "info": {
    "title": "Sofa API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "{protocol}://{host}:{port}/api/v1",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "3000"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "name": "better-auth.session_token",
        "in": "cookie",
        "description": "Better Auth session cookie"
      }
    },
...
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/admin/backups/schedule": {
      "get": {
        "operationId": "admin.backups.schedule",
        "summary": "Get backup schedule",
        "description": "Fetch the current automated backup schedule configuration.",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Backup schedule settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean",
                      "description": "Whether scheduled backups are enabled"
                    },
                    "maxRetention": {
                      "type": "number",
                      "description": "Maximum backups to keep (0 = unlimited)"
                    },
                    "frequency": {
                      "enum": [
                        "6h",
                        "12h",
                        "1d",
                        "7d"
                      ],
                      "description": "Backup interval"
                    },
                    "time": {
                      "type": "string",
                      "description": "Scheduled time (HH:MM, 24-hour format)"
                    },
                    "dayOfWeek": {
                      "type": "number",
                      "description": "Day of week for weekly backups (0 = Sunday)"
                    }
                  },
                  "required": [
                    "enabled",
                    "maxRetention",
                    "frequency",
                    "time",
                    "dayOfWeek"
                  ],
                  "description": "Automated backup schedule configuration"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "operationId": "admin.backups.updateSchedule",
        "summary": "Update backup schedule",
        "description": "Update the automated backup schedule. Only provided fields are changed; omitted fields keep their current values.",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean",
                    "description": "Enable or disable scheduled backups"
                  },
                  "frequency": {
                    "enum": [
                      "6h",
                      "12h",
                      "1d",
                      "7d"
                    ],
                    "description": "Backup interval"
                  },
                  "time": {
                    "type": "string",
                    "pattern": "^\\d{2}:\\d{2}$",
                    "description": "Time of day to run backups (HH:MM, 24-hour format)"
                  },
                  "dayOfWeek": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 6,
                    "description": "Day of week for weekly backups (0 = Sunday, 6 = Saturday)"
                  },
                  "maxRetention": {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991,
                    "description": "Maximum number of backups to keep (0 = unlimited, 1-30 otherwise)"
                  }
                },
                "description": "Partial update to the automated backup schedule"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/admin/telemetry": {
      "get": {
        "operationId": "admin.telemetry",
        "summary": "Get telemetry status",
        "description": "Fetch whether anonymous telemetry is enabled and when the last report was sent.",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Telemetry configuration and last report time",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean",
                      "description": "Whether telemetry is currently enabled"
                    },
                    "lastReportedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "ISO 8601 timestamp of the last telemetry report, or null if never sent"
                    }
                  },
                  "required": [
                    "enabled",
                    "lastReportedAt"
                  ],
                  "description": "Telemetry configuration and last report time"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "operationId": "admin.toggleTelemetry",
        "summary": "Toggle telemetry",
        "description": "Enable or disable anonymous telemetry reporting.",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean",
                    "description": "Whether anonymous telemetry reporting is enabled"
                  }
                },
                "required": [
                  "enabled"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  }
}